# Conflicts:
#	gradle.properties
#	omniNotes/build.gradle
#	omniNotes/src/main/java/it/feio/android/omninotes/async/notes/SaveNoteTask.java

+++ omniNotes/src/main/java/it/feio/android/omninotes/ListFragment.java
@@ -626,49 +626,52 @@ public class ListFragment extends BaseFragment implements OnViewTouchedListener,
 
         MenuItemCompat.setOnActionExpandListener(searchMenuItem, new MenuItemCompat.OnActionExpandListener() {
 
-            boolean searchPerformed = false;
+			boolean searchPerformed = false;
 
 
-            @Override
-            public boolean onMenuItemActionCollapse(MenuItem item) {
-                // Reinitialize notes list to all notes when search is collapsed
-                searchQuery = null;
-                if (searchLayout.getVisibility() == View.VISIBLE) {
-                    toggleSearchLabel(false);
-                }
-                mainActivity.getIntent().setAction(Intent.ACTION_MAIN);
-                initNotesList(mainActivity.getIntent());
-                mainActivity.supportInvalidateOptionsMenu();
-                return true;
-            }
+			@Override
+			public boolean onMenuItemActionCollapse(MenuItem item) {
+				// Reinitialize notes list to all notes when search is collapsed
+				searchQuery = null;
+				if (searchLayout.getVisibility() == View.VISIBLE) {
+					toggleSearchLabel(false);
+				}
+				mainActivity.getIntent().setAction(Intent.ACTION_MAIN);
+				initNotesList(mainActivity.getIntent());
+				mainActivity.supportInvalidateOptionsMenu();
+				return true;
+			}
 
 
-            @Override
-            public boolean onMenuItemActionExpand(MenuItem item) {
-                searchView.setOnQueryTextListener(new OnQueryTextListener() {
-                    @Override
-                    public boolean onQueryTextSubmit(String arg0) {
-                        return prefs.getBoolean("settings_instant_search", false);
-                    }
+			@Override
+			public boolean onMenuItemActionExpand(MenuItem item) {
+
+				searchView.setOnQueryTextListener(new OnQueryTextListener() {
+					@Override
+					public boolean onQueryTextSubmit(String arg0) {
+
+						return prefs.getBoolean("settings_instant_search", false);
+					}
 
 
-                    @Override
-                    public boolean onQueryTextChange(String pattern) {
-                        if (prefs.getBoolean("settings_instant_search", false) && searchLayout != null &&
-                                searchPerformed && mFragment.isAdded()) {
-                            searchTags = null;
-                            searchQuery = pattern;
-                            NoteLoaderTask.getInstance().execute("getNotesByPattern", pattern);
-                            return true;
-                        } else {
-                            searchPerformed = true;
-                            return false;
-                        }
-                    }
-                });
-                return true;
-            }
-        });
+					@Override
+					public boolean onQueryTextChange(String pattern) {
+
+						if (prefs.getBoolean("settings_instant_search", false) && searchLayout != null &&
+								searchPerformed && mFragment.isAdded()) {
+							searchTags = null;
+							searchQuery = pattern;
+							NoteLoaderTask.getInstance().execute("getNotesByPattern", pattern);
+							return true;
+						} else {
+							searchPerformed = true;
+							return false;
+						}
+					}
+				});
+				return true;
+			}
+		});
     }
 
 
@@ -840,15 +843,15 @@ public class ListFragment extends BaseFragment implements OnViewTouchedListener,
     void editNote(final Note note, final View view) {
         if (note.isLocked() && !prefs.getBoolean("settings_password_access", false)) {
             BaseActivity.requestPassword(mainActivity, passwordConfirmed -> {
-                if (passwordConfirmed) {
-                    note.setPasswordChecked(true);
-                    AnimationsHelper.zoomListItem(mainActivity, view, getZoomListItemView(view, note),
-                            listRoot, buildAnimatorListenerAdapter(note));
-                }
-            });
+				if (passwordConfirmed) {
+					note.setPasswordChecked(true);
+					AnimationsHelper.zoomListItem(mainActivity, view, getZoomListItemView(view, note),
+							listRoot, buildAnimatorListenerAdapter(note));
+				}
+			});
         } else {
             AnimationsHelper.zoomListItem(mainActivity, view, getZoomListItemView(view, note),
-                    listRoot, buildAnimatorListenerAdapter(note));
+					listRoot, buildAnimatorListenerAdapter(note));
         }
     }
 
@@ -944,14 +947,26 @@ public class ListFragment extends BaseFragment implements OnViewTouchedListener,
                 .content(R.string.empty_trash_confirmation)
                 .positiveText(R.string.ok)
                 .callback(new MaterialDialog.ButtonCallback() {
-                    @Override
-                    public void onPositive(MaterialDialog materialDialog) {
-                        for (int i = 0; i < listAdapter.getCount(); i++) {
-                            selectedNotes.add(listAdapter.getItem(i));
-                        }
-                        deleteNotesExecute();
-                    }
-                }).build().show();
+					@Override
+					public void onPositive(MaterialDialog materialDialog) {
+
+						boolean mustDeleteLockedNotes = false;
+						for (int i = 0; i < listAdapter.getCount(); i++) {
+							selectedNotes.add(listAdapter.getItem(i));
+							mustDeleteLockedNotes = mustDeleteLockedNotes || listAdapter.getItem(i).isLocked();
+						}
+						if (mustDeleteLockedNotes) {
+							mainActivity.requestPassword(mainActivity, getSelectedNotes(),
+									passwordConfirmed -> {
+										if (passwordConfirmed) {
+											deleteNotesExecute();
+										}
+									});
+						} else {
+							deleteNotesExecute();
+						}
+					}
+				}).build().show();
     }
